Update hypothesis in poetry.lock file
In test_repr, we need to import data first otherwise the new classes
will conflict with the test
In test_data, we avoid a warning by using float instead of np.float

+++ tests/raw/test_data.py
@@ -35,7 +35,7 @@ def draw_test_data(draw, shape):
     if len(shape) == 0:
         result = draw(elements)
     else:
-        result = draw(np_strat.arrays(np.float, shape, elements=elements))
+        result = draw(np_strat.arrays(float, shape, elements=elements))
     return np.sign(result) * np.maximum(np.abs(result), 1 / threshold)
